requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
bfg==0.1.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
distro==1.9.0
Django==5.1.5
django-crispy-forms==2.3
djangorestframework==3.15.2
firebase-admin==6.6.0
frozenlist==1.6.0
gcloud==0.18.3
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyper==0.7.0
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
jwcrypto==1.5.6
msgpack==1.1.0
multidict==6.4.4
numpy==2.3.0
oauth2client==4.1.3
openai==1.86.0
packaging==25.0
pandas==2.3.0
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.1
Pyrebase4==4.8.0
pyTelegramBotAPI==4.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
python-telegram-bot==20.3
pytoml==0.1.21
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
whitenoise==6.9.0
yarl==1.20.0
